requirements.txt
opencv-python
numpy
pyttsx3
# tflite-runtime # Use this on RPi
tensorflow-intel; platform_system == "Windows"
RPi.GPIO; platform_system == "Linux"
picamera; platform_system == "Linux"
